improvements after code review
