+++ backend/src-common/src/main/resources/sw360.properties
@@ -36,6 +36,7 @@ subjectForNewModerationRequest= New moderation request
 subjectForUpdateModerationRequest= Update on moderation request
 subjectForAcceptedModerationRequest= Your moderation request has been accepted
 subjectForDeclinedModerationRequest= Your moderation request has been declined
+subjectForDeclinedUserModerationRequest= Your request for SW360 user account has been declined
 
 textForNewModerationRequest= a new moderation request has been added to your SW360-account.\n\n
 textForUpdateModerationRequest= \
@@ -43,3 +44,4 @@ textForUpdateModerationRequest= \
   SW360-account has been updated.\n\n
 textForAcceptedModerationRequest= your moderation request has been accepted by one of the moderators.\n\n
 textForDeclinedModerationRequest= your moderation request has been declined by one of the moderators.\n\n
+textForDeclinedUserModerationRequest= your request for SW360 user account has been declined by one of the moderators.\n\n
