+++ backend/src/src-moderation/src/main/java/com/siemens/sw360/moderation/db/ModerationDatabaseHandler.java
@@ -27,7 +27,6 @@ import com.siemens.sw360.datahandler.thrift.RequestStatus;
 import com.siemens.sw360.datahandler.thrift.components.Component;
 import com.siemens.sw360.datahandler.thrift.components.Release;
 import com.siemens.sw360.datahandler.thrift.licenses.License;
-import com.siemens.sw360.datahandler.thrift.licenses.Todo;
 import com.siemens.sw360.datahandler.thrift.moderation.DocumentType;
 import com.siemens.sw360.datahandler.thrift.moderation.ModerationRequest;
 import com.siemens.sw360.datahandler.thrift.projects.Project;
@@ -37,7 +36,6 @@ import com.siemens.sw360.datahandler.thrift.users.UserService;
 import com.siemens.sw360.datahandler.permissions.PermissionUtils;
 import com.siemens.sw360.datahandler.thrift.ThriftClients;
 import com.siemens.sw360.licenses.db.LicenseDatabaseHandler;
-import com.siemens.sw360.licenses.db.LicenseTypeRepository;
 import com.siemens.sw360.mail.MailConstants;
 import com.siemens.sw360.mail.MailUtil;
 import org.apache.log4j.Logger;
@@ -139,7 +137,7 @@ public class ModerationDatabaseHandler {
         ModerationRequest request = repository.get(requestId);
         request.moderationState = ModerationState.REJECTED;
         repository.update(request);
-        sendMailToUserForDeclinedRequest(request.getRequestingUser());
+        sendMailToUserForDeclinedRequest(request.getRequestingUser(), request.getDocumentType() == DocumentType.USER);
     }
 
 
@@ -325,9 +323,13 @@ public class ModerationDatabaseHandler {
         mailUtil.sendMail(moderators,MailConstants.SUBJECT_FOR_UPDATE_MODERATION_REQUEST,MailConstants.TEXT_FOR_UPDATE_MODERATION_REQUEST);
     }
 
-    private void sendMailToUserForDeclinedRequest(String userEmail){
+    private void sendMailToUserForDeclinedRequest(String userEmail, boolean userRequest){
         MailUtil mailUtil = new MailUtil();
-        mailUtil.sendMail(userEmail,MailConstants.SUBJECT_FOR_DECLINED_MODERATION_REQUEST,MailConstants.TEXT_FOR_DECLINED_MODERATION_REQUEST);
+        if (userRequest){
+            mailUtil.sendMail(userEmail, MailConstants.SUBJECT_FOR_DECLINED_USER_MODERATION_REQUEST, MailConstants.TEXT_FOR_DECLINED_USER_MODERATION_REQUEST, false);
+        } else {
+            mailUtil.sendMail(userEmail, MailConstants.SUBJECT_FOR_DECLINED_MODERATION_REQUEST, MailConstants.TEXT_FOR_DECLINED_MODERATION_REQUEST, true);
+        }
     }
