+++ frontend/sw360-portlet/src/main/webapp/html/moderation/users/merge.jsp
@@ -1,5 +1,5 @@
 <%--
-  ~ Copyright Siemens AG, 2013-2015. Part of the SW360 Portal User.
+  ~ Copyright Siemens AG, 2013-2016. Part of the SW360 Portal User.
   ~
   ~ This program is free software; you can redistribute it and/or modify it under
   ~ the terms of the GNU General Public License Version 2.0 as published by the
